Gemini API/1 CSV Query/requirements.txt
faiss-cpu 
google-generativeai
langchain_google_genai
langchain
langchain-community
python-dotenv
streamlit